feat(hal): port i2c from STM32F1xx to LS2K03xx
Remove dma, 10bit address and dual address.

+++ Drivers/LS2K03xx_HAL_Driver/Inc/ls2k03xx_hal.h
@@ -95,7 +95,7 @@ void HAL_MspDeInit(void);
   */
 /* Peripheral Control functions  ************************************************/
 void HAL_Delay(uint32_t Delay);
-uint64_t HAL_GetTick(void);
+uint32_t HAL_GetTick(void);
 /**
   * @}
   */
